ENG-715: Use shared locks for YQL pure-write requests.
Summary:
When running workload A of YCSB benchmark, there is a high lock contention resulting in low IOPs. The workload A consists of 50% read/write. The default key distribution is Zipfian in which 80% of requests are distributed among 20% of the keys (i.e. hot-key distribution). Currently, YugaByte attempts to acquire an exclusive (snapshot) lock on the row being written unconditionally, which results in the lock contention.

This diff changes `PrepareDocWriteOperation()` to acquire a shared (serializable) lock on the row if the write request does not require a read snapshot. Refer to <doc-link> for lock compatibility matrix.

This diff also fixes a bug in SQL analyzer that erroneously add columns being written for DMLs even when they do not require a read snapshot (i.e. pure-write that does not involve IF condition nor counter column update).

Test Plan: Ran YCSB workload A (50/50 read/write) with Zipfian (hot-key) distribution and verified no IOPs reduction and tserver call latency is below 1 ms (was 10's of ms with row-level locks).

Reviewers: mikhail, neil, karthik, kannan, mihnea, sergei

Reviewed By: mihnea, sergei

Subscribers: akashnil, sergei, ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D3094