ENG-2191 Add user timestamp to the value portion in DocDB
Summary:
This diff adds an optional user timestamp specified in microseconds to the value portion of
DocDB. I've changed our current code to ensure we always skip this field using
DecodeWithoutUserTimestamp. As we implement the core docdb functionality for USING TIMESTAMP, we'll
make use of the timestamp field in appropriate places.

Test Plan:
1) unit tests
2) jenkins

Reviewers: akashnil, mikhail

Reviewed By: mikhail

Subscribers: sergei, ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D3180