ENG-2241 Introduce REDIS commands TSADD and TSGET for redis timeseries datatype.
Summary:
The timeseries datatype is similar to hash maps in REDIS, with the key difference being
that the key is an int64 and is sorted in descending order to better support a timeseries like
usecase.

In this diff we introduce the basic operations TSADD (to add a timestamp value pair) and TSGET (to
retrieve the value for a particular timestamp).

Test Plan:
1) Add tests
2) jenkins

Reviewers: sergei, hector, amitanand, akashnil

Reviewed By: akashnil

Subscribers: yql, ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D3225