ENG-2252: Fix for spot price editing bug
Summary: As part of a recent change, normalization of spot price started happening on every edit when it was only supposed to happen when finished editing (on focus change). This fixes that while maintaining assurance that invalid values will not be sent to the backend while fetching universe resources.

Test Plan: Go to UI, validate bug no longer appears.

Reviewers: pritam

Reviewed By: pritam

Subscribers: jenkins-bot, kannan, ui

Differential Revision: https://phabricator.dev.yugabyte.com/D3228