Adding customer email in navbar
Summary:
With our recent multi-tenant changes, it's becoming important to see which customer account
we're actually in. Added a small entry to the navbar next to user profile dropdown to showcase the
email address of the logged in user.

Test Plan: tested from UI: https://phabricator.dev.yugabyte.com/F5738

Reviewers: jeff, bharat, ram, pritam

Reviewed By: pritam

Subscribers: jenkins-bot, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D3229