ENG-2209: Show SST file sizes per tablet in the metrics page.
Summary: This change adds the total size of all the SST files and displays them on the Yugabyte UI's metrics page.

Test Plan:
Tested it by running various workloads and then manually testing that the size displayed on the metrics page is correct
by looking at all the files that were written on disk.

Reviewers: timur, mikhail, pritam.damania

Reviewed By: pritam.damania

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D3263