ENG-2276 Ensure we handle min max ttl values correctly.
Summary:
kMaxTtl was set to int32_t::max() and was used in DocDB to indicate a missing/invalid TTL
value. This was originally intended to be the max TTL value for Cassandra, but was repurposed in
DocDB as well. This would cause an issue that if set a TTL of int32_t::max() in Cassandra, the TTL
value would be ignored.

With this diff, I've decoupled this and now kMaxTtl denotes the special missing/invalid TTL value
only in DocDB and is set to int64_t::max(). In addition to this, we have separate min and max values
just for Cassandra.

Test Plan:
1) unit tests added which were failing before this diff.
2) jenkins.

Reviewers: mihnea, robert, neil

Reviewed By: neil

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D3275