ENG-2222: Enhance yb_release to actually build community
Summary:
Previously I hadn't fully tested that the community edition was being built if we were
passing the respective community edition flag, as the default would build enterprise and that is
what the default release does from devops.

Glued together the flag to the build.

Test Plan: tested with local yb_release as well as full path from itest locally.

Reviewers: mikhail, bharat

Reviewed By: bharat

Subscribers: devops, ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D3281