Don't crash on failure to acquire a leader lease before loading system catalog data if term changed
Summary:
This is for ENG-2268.

Problem:

The master became a leader, started loading system catalog, and then reverted to a follower. The
system catalog loading task failed to acquire a leader lease and crashed.

The reason this only started happening with leader leases is probably the fact that previously we
were only waiting for all pending operations to be committed, which would normally always be quick
even in the follower case if there is no network partition. Even if a master started with
uncommitted operations in the log, became a leader, then reverted to the follower with those
operations still uncommitted and system catalog load task in progress, the new master leader would
quickly commit those old records or overwrite them, unblocking the task. Waiting for leader lease,
however, never succeeds if we stop being the leader.

Solution:
Don't crash if we fail when waiting for all changes to be applied, but the term has changed anyway.

Test Plan: Jenkins

Reviewers: bogdan, bharat

Reviewed By: bharat

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D3302