[ENG-2472] Fix type-casting issue when incrementing counters with negative numbers
Summary:
This adds the conversion methods from varint to the other numeric types since we parse integer
literals as varints.

For constants this is not needed because they are directly evaluated to the expected type.
But we treat negative numbers as expressions using the unary minus operator, so we need the
implicit conversion in that case (e.g. '.. SET c = c + -1 ..')

Test Plan: updated TestArith.java

Reviewers: neil, robert

Reviewed By: robert

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D3591