Remove column id in QL protobuf.
Summary:
Given that the primary key values in QLReadRequestPB and QLWriteRequestPB are required to be in the same order as the primary key columns in the table schema, the column IDs in those values in the protobuf should be removed because they are redundant.

The column IDs are safe to remove becaue the request PB carries the table schema version that is used to ensure the request is built using the same table schema that the tablet is using.

Test Plan: YB CQL C++ and Java tests

Reviewers: neil, mihnea, bogdan

Reviewed By: mihnea

Subscribers: karthik, yql

Differential Revision: https://phabricator.dev.yugabyte.com/D3327