ENG-2304: Fix for configure call not getting sent
Summary: Root cause: We were setting the initial spot price to '0.0' instead of '0.00' which is what float normalization provides. Quick and simple fix is to change initial spot price to '0.00'. Reason it was only an issue in production is that in development, we default to using spot price and start in a non-0.0 state.

Test Plan: Go to ui, change to production state (set IN_DEVELOPMENT_MODE to false), validate placement info appears. Repeat for development state.

Reviewers: pritam, bogdan, ram

Reviewed By: ram

Subscribers: jenkins-bot, ui

Differential Revision: https://phabricator.dev.yugabyte.com/D3330