ENG-2322: Toggling spot price on and off is not reflected in universe cost in UI
Summary: Need to reset spot price to 0.00 when toggled off. Because we fetch a suggested spot price when we toggle on, there's no need to persist the previous spot price.

Test Plan: Go to UI Create Universe page, toggle off and back on and confirm that the bug no longer appears

Reviewers: ram, pritam

Reviewed By: pritam

Subscribers: jenkins-bot, kannan, ui

Differential Revision: https://phabricator.dev.yugabyte.com/D3347