ENG-2321: Fix for spot price change in edit universe bug
Summary:
Spoke with @bogdan and @karthik and we decided we do not want to support changes to spot price in the Edit Universe form at the moment. This is because spot price changes should be a full move and there are several things which can cause a spot price change. We were already intending to not support spot price changes in Edit Universe, but did not realize that some of edits we do support were causing spot price changes.

The simple fix for now has 2 parts:
1) disable changing instance type for a spot-price-based universe.
2) ignore requests to get a suggested spot price in the edit universe flow (this is needed if, for example, a customer adds another region)

Test Plan:
- Go to UI -> create a spot-price based universe -> go to edit universe form -> verify instance type is disabled and changes to region do not cause spot price change
- In same UI session, verify non-spot price based universe allows editing instance type
{F6032}

Reviewers: pritam, ram, bogdan

Reviewed By: bogdan

Subscribers: jenkins-bot, bogdan, karthik, ui

Differential Revision: https://phabricator.dev.yugabyte.com/D3356