Basic java framework for running unit tests with the Jedis client against YB.
Summary:
This diff introduces the basic framework for writing unit tests in Java which use the Jedis
client pointed to our YB minicluster. This is similar to how we have Java tests for the Cassandra
drivers.

Also, a very simple unit test has been added to demonstrate this works.

Test Plan:
1) jenkins
2) unit test added.

Reviewers: mikhail, hector

Reviewed By: hector

Subscribers: kannan, ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D3372