ENG-2191 Support 'USING TIMESTAMP' for Deletes.
Summary:
This diff introduces support for 'USING TIMESTAMP' in DELETE statements for CQL. Deletes
are similar to inserts in terms of how they are handled in docdb with the 'USING TIMESTAMP' clause.
An important distinction is that in CQL, we support full row deletes. In this particular, case we
actually write a tombstone for each individual column in DocDB instead of writing a tombstone at the
row level. The reason for this is that a row level DELETE with a timestamp might only end up
deleting some columns of the row and not others (which are at higher timestamps). Since, we don't
want a read penalty for 'USING TIMESTAMP', we don't want to resolve this during reads. As a result,
we write a tombstone for every possible column in the schema.

Test Plan:
1) unit tests.
2) jenkins

Reviewers: mikhail, robert, akashnil

Reviewed By: akashnil

Subscribers: neil, ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D3411