Implement CREATE INDEX syntax.
Summary:
The syntax is:

```CREATE INDEX [IF NOT EXISTS] index_name
ON [keyspace_name.]table_name ( (index_hash_column, ...), index_cluster_column, ... )
[WITH CLUSTERING ORDER BY (index_cluster_column { ASC | DESC }, ...)]
[COVERING (covered_column, ...)]```

The `index_hash_column` and `index_cluster_column` specify how the indexed column(s) will be hashed / clustered in the index. They are the primary key of the index and can be any columns from the indexed table. The `covered_column` specifies the additional column(s) in the indexed table that will be covered in the index.

- Only the parser and semantic analyzer have been implemented. Will raise "FEATURE NOT SUPPORTED" error if CREATE INDEX is attempted currently.

Test Plan: ql-parser-test and ql-semanalyzer-test for the new syntax.

Reviewers: neil, pritam.damania

Reviewed By: pritam.damania

Subscribers: karthik, yql

Differential Revision: https://phabricator.dev.yugabyte.com/D3200