ENG-2235 Ensure collection's TTL does not override individual element's TTL in DocDB.
Summary:
In cassandra if an entire collection was written with say TTL 10 and then a single element
was overwritten/added with TTL 20, then when the TTL of 10 expires, the single element still
survives.

Our implementation for collections wasn't consistent with this since in DocDB the init
marker for a collection was considered as a tombstone at the TTL expiry time. As a result, this
would hide all values written at a time lower than the expiry time. We've fixed this issue in this
diff to ensure that the TTL for the init marker only applies to itself and not the entire
subdocument. The same also applies for compactions.

Test Plan:
1) unit tests.
2) jenkins

Reviewers: akashnil, mikhail

Reviewed By: mikhail

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D3457