Enhance version mismatch message for edit from stale page.
Summary:
If create was in process, and version is older, then edit would run into a `user version is different from latest version` scenario. This change enhances the error message to suggest that it is expected and user has to retry from a refreshed page.

Test Plan: Just a error/log message change.

Reviewers: karthik, jeff

Reviewed By: jeff

Subscribers: jenkins-bot, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D3442