ENG-2370: Fix intermittent wrong placement_region in TestSystemTables.testSystemLocalTables
Summary:
`select data_center from system.local` returns different placement_region intermittently than the one specified when the cluster is created. This happens because the 4th tserver started in the test is not started with the specified placement_region. Therefore "data_center" column value can be different depending on which node the query is routed to.

Fixing to pass tserver args with placement_region also when starting the new yb-tserver.

Test Plan: Run org.yb.cql.TestSystemTables and verify in org.yb.cql.TestSystemTables-output.txt that the 4th yb-tserver is started with the correct placement_region.

Reviewers: pritam.damania

Reviewed By: pritam.damania

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D3447