Redis SET Update/Insert semantics. Avoid look up for upsert operations.
Summary:
Add support for SET [NX|XX].

For SET [Upsert] operations, there is no need to look up the type of the existing key. This is needed only for inserts into the collection type viz: {hash/TS/Set} or when [NX|XX] is specified.
Having an unnecessary look up causes an extra seek, that can be totally avoided in the write path.

Ensure that the redis proxy specifies REDIS_TYPE_STRING as the type instead of NONE, as previously done. This change will cause SET requests to fail, if an "old" redis proxy is talking to a "new" Tserver. Something to watch out for, during rolling restarts.

Test Plan:
Jenkins for any possible regressions.
Push and check metrics for perf impact.

Set Latency graph. Before and after the update. (Updated around 14:15)
https://phabricator.dev.yugabyte.com/F6598

Reviewers: akashnil, pritam.damania, kannan

Reviewed By: pritam.damania

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D3458