Make pg_version_num() consistent with sever_version_num.
Unless building on 8.2 or earlier. Includes an extension upgrade script that
updates the function, too. Also tweak the test description to reflecta that
it's the `_num` variant that's being tested.

Patch changes are to get line numbers re-aligned, as well as to go back
to the old version of `pg_version_num()` on 8.2.

Resolves #148.