ENG-2398: fix for form validation bug with default suggested spot price
Summary: As in the JIRA, using the default suggested spot price causes validation to fail on the Create Universe form. Root cause is that the value of the component was being set, but not the value of the redux form field. Fix is to move the code that sets the initial value to componentWillMount from render and to use the field's built in `onChange` function to set the value.

Test Plan:
Go to UI, confirm bug no longer appears on form submit.
Create spot price universe, confirm it is created with the correct spot price.

Reviewers: pritam

Reviewed By: pritam

Subscribers: jenkins-bot, ui

Differential Revision: https://phabricator.dev.yugabyte.com/D3498