Fix locking issue on /table endpoint in YB master.
Summary:
In master path handlers, `CallIfLeaderOrPrintRedirect` was already taking the catalog
manager leader shared lock and then was invoking `HandleTablePage` for the /table endpoint.
`HandleTablePage` also tried to acquire the same lock and as a result there was a CHECK failure.

In this diff, I've added a unit test that reproduces the issue and also fixed the issue by removing
the lock from HandleTablePage.

Test Plan:
1) unit tests
2) jenkins

Reviewers: bharat

Reviewed By: bharat

Subscribers: ybase, bogdan

Differential Revision: https://phabricator.dev.yugabyte.com/D3502