[ENG-2376] Fix overflow issue with timestamp comparison
Summary:
For timestamp values we use microsecond precision internally but interpret numeric inputs as
milliseconds. This can cause an overflow for very large or small values (well outside the
supported range for formatted dates: years 1400 to 9999).

In particular setting max/min Int64 as upper/lower bounds for a range query on timestamp might
not return all results due to overflow.
We handle overflow by returning max/min value in microsecond in that case so that comparisons
work as expected.

Test Plan: TestTimestampDataType

Reviewers: robert

Reviewed By: robert

Subscribers: bharat, yql

Differential Revision: https://phabricator.dev.yugabyte.com/D3512