ENG-2380 Ensure leader TServer and Master handle WRONG_SERVER_UUID correctly.
Summary:
If we stop a tserver, wipe its data and bring it back up on the same host and port (within
5 mins - which is the timeout after which a leader evicts a follower), it starts up with a new UUID.
As a result, the tservers contacting it as part of its old quorum would get a WRONG_SERVER_UUID
error. Although, the tservers didn't handle this error correctly and would indefinitely try to
replicate OPs to this tserver.

We had the same issue for the master as well, where we would try to delete the tablet on this
tserver, get WRONG_SERVER_UUID and indefinitely retry.

Apart from this, another issue was that YBClient in the proxies did not handle TABLET_NOT_FOUND
correctly and would continously try to read/write from the tserver that was wiped out.

In this diff, I've fixed the above issues as follows:
1. For the tserver, we evict the follower once we see WRONG_SERVER_UUID.
2) For the master, we stop retrying the delete request once we see WRONG_SERVER_UUID.
3) In YBClient, we had TABLET_NOT_FOUND correctly and fail over to a different server.

Test Plan:
1) unit tests.
2) jenkins

Reviewers: bogdan, robert, bharat

Reviewed By: bharat

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D3524