[ENG-1894] Support Range Deletes in YQL
Summary:
Allow deletes where range portion of primary key is not fully specified.
Then all rows in that hash that match the where clause will be deleted.

Currently, this comes up as a corner case in an application deleting old data that is not being concurrently modified.
Therefore, the current implementation is optimized for performance so the deletes are executed sequentially not as a transaction.
This will be addressed in a followup commit.

Test Plan: Updated java TestDelete

Reviewers: neil, akashnil, robert

Reviewed By: robert

Subscribers: yql, bharat

Differential Revision: https://phabricator.dev.yugabyte.com/D3541