ENG-2422: Aggregating back metrics split per device and per table
Summary:
After switching to node + YB metrics instead of collectd, we were left with a set of
metrics that were reported differently than desired. In particular, disk info was reported as
averages across the devices and the new per table level metrics from YB were reported as averages
across all tables.

This diff fixes that while still staying in the confines of our current yml processing framework,
through the use of the "without" keyword in Prometheus, that allows for aggregations that would
ignore certain labels. In this case, we mark the labels which were forcing breakdowns (device for
disk metrics and table_id and table_name for table metrics) and as such, manage to sum across those
entities.

Test Plan:
Ran a workload and tested the queries in the prometheus graph endpoint directly as well
to match. Added a different Cassandra workload afterwards and confirmed the graphs didn't all take
a big dive.

Reviewers: ram, bharat, jeff, amitanand

Reviewed By: amitanand

Subscribers: amitanand, jenkins-bot, karthik, kannan, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D3537