ENG-2423: Remove options for num tablets/replicas from yb-admin setup_redis_table
Summary:
Removing flags for setting up num_tablets and num_replicas so we leave CreateTable to
operate on default behavior. Right now, this does the following:
- for RF it uses the universe config, if available (YW) else replication_factor flag (set by yb-ctl)
- for num_tablets, we default to flag yb_num_shards_per_tserver times number of live tservers

There is still the small caveat that one might run `yb-ctl setup_redis` before all the tservers
have reported to the master leader, but for that we need mechanics to wait for a master leader and
for a given number of tablet servers.

Test Plan:
ran yb-ctl with --replication_factor 1, 3 and 5 and confirmed 8, 24 and 40 respective
tablets with yb-admin list_tablets

Reviewers: bharat, karthik, kannan

Reviewed By: kannan

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D3543