ENG-2444: Periodic syncs when using buffered WAL
Summary:
When durable_wal_write is false, we currently do buffered WAL IO (rather than using O_DIRECT).
In this mode, it is desirable if we could do some periodic syncs on the WAL to minimize amount of
potential loss.
We now provide  GFLAGs to sync the WAL when either one of the following conditions happen:
every configurable number of millisecs (say default 1000ms) when there's unsynced data in the WAL
OR, when there's some configurable amount (say default 1MB) of unsync'ed data in the WAL

Test Plan: Added tests TestFsyncInterval and TestFsyncDataSize in LogTest.

Reviewers: mikhail, sergei, kannan

Reviewed By: sergei, kannan

Subscribers: hector, sergei, ybase, bharat

Differential Revision: https://phabricator.dev.yugabyte.com/D3554