ENG-2418: CQL consistency level ONE should round robin between replicas in same DC.
Summary: This code resolves the error that when the CQL consistency level is set to one, the reads ends up going only to the leader. We resolve this error by shuffling the list of hosts, because by default the first host in the list is always the leader.

Test Plan:
Tested it by:

1) Writing a unit test that checks that the operations are roughly equivalently distributed accross the replicas given a standard deviation factor of 2. In short, each replicas get atleast (total / #number of replicas * deviation factor) number of ops and the difference between the replicas revieving the maximum and the minimum number of ops is not more than that.
2) By manually verifying by running the test for various number of ops and checking that the number of ops are almost equivalently distributed accross the replicas.

Reviewers: karthik, sid, pritam.damania, robert

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D3568