ENG-2451 Skip entry in system.peers table if there is an issue resolving the IP.
Summary:
While building entries for system.peers, we would perform a DNS lookup if a hostname was
provided in the registration. If this lookup failed, we would fail the entire SELECT operation. This
wasn't ideal and hence I've changed the behavior to skip such entires in case we couldn't lookup the
appropriate IP address.

Test Plan:
1) unit test that reproduces the issue.
2) jenkins

Reviewers: mihnea, robert

Reviewed By: robert

Subscribers: ram, sid, bogdan, yql, bharat

Differential Revision: https://phabricator.dev.yugabyte.com/D3565