ENG-2392: consensus: always send at least one op
Summary:
This fixes a bug in which the LogCache would not return any operations
in the case of a cache miss on an operation larger than the configured
batch size.

This would cause consensus to get "stuck": the leader would repeatedly
send status-only messages (i.e. no operations) to the follower in a
tight loop, never making any progress.

The new unit test reproduces the problem.

Ported change from Kudu.
https://github.com/apache/kudu/commit/5ba18b623095170b55c2bab90a09258853cce1b8

Test Plan: test and jenkins

Reviewers: venkatesh, mikhail

Reviewed By: mikhail

Subscribers: ybase, bharat

Differential Revision: https://phabricator.dev.yugabyte.com/D3574