ENG-2380 Ensure master keeps only a single valid node for a particular host and port.
Summary:
Currently, the master's map for tservers is keyed on the uuid. As a result, we could have
two tservers with same host and port but different uuids be stored in the master's map. This happens
when we wipe a tservers data directory and bring it back up. This scenario causes inconsistent views
on the master's webpage since both tservers with the same host and port show up in several areas. In
addition to this, any code that reads this map would end up running into some sort of inconsistent
behavior due to this.

To fix this issue, I've added a `removed` state in the TSDescriptor to indicate that this tserver
should be ignored. I didn't remove the entry from the map since several places held references as
naked pointers to this object and those areas would run into segfaults as a result.

Test Plan:
1) unit tests.
2) jenkins

Reviewers: oleg, bharat

Reviewed By: bharat

Subscribers: ybase, bogdan

Differential Revision: https://phabricator.dev.yugabyte.com/D3583