ENG-2015 Improve fault tolerance for creating system tables.
Summary:
We ran into an issue where while creating a system table, the tablet was created but the
associated table wasn't created due to a leader election. As a result, the new leader would find the
tablets but couldn't find the table which would result in a FATAL crash.

To ensure this is handled correctly, we first create the table and then the associated tablets. In
addition to this, if we find that a particular table was created, but the tablets are missing, we
still re-create the table.

Test Plan:
1) Run MasterChangeConfigTest.TestNewLeaderWithPendingConfigLoadsSysCatalog 100 times.
This test was failing intermittently due to this issue earlier.
2) jenkins

Reviewers: bogdan, oleg, bharat

Reviewed By: bharat

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D3596