ENG-2500, Remove Node button missing in Node Details table
Summary:
In our earlier design when node is in ToBeAdded state we show a corresponding Remove Node button.
That logic broke because of changes to React-Boostrap-Table invocation and also due to changing the Status string.

Put fixes for the same.

Test Plan:
Create a universe.
Go to NodeDetails page.
You should see the remove node button while the node is in "ToBeAdded" state.

Click on it. The node should be removed from the universe details json.

Reviewers: ram, jeff

Reviewed By: jeff

Subscribers: jenkins-bot, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D3609