ENG-2282 YW UniverseDefinitionTaskParams clone is wrong
Summary:
Fixed clone for Objects.

Strings are immutable so just assigning the reference should be fine.

Test Plan:
The clone method should not alter the objects it's cloning.
We currently do not use it in any way that affects the outcome (Eg. even though the original object gets modified we are not persisting in db) until a final task is fired.

Reviewers: ram, bogdan, jeff

Reviewed By: jeff

Subscribers: jeff, jenkins-bot, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D3339