Make the table details page have the table name in the URL
Summary: Make the table details page have the table name in the URL. Currently it has the table uuid in the url which makes it hard to create a stable link. It would be good to have an alternate path where we can browse to the table details using the URL directly.

Test Plan:
1) Manually tested that it works by looking at the UI page.
2) Jenkins.

Reviewers: karthik, pritam.damania, bogdan

Reviewed By: bogdan

Subscribers: ybase, bharat

Differential Revision: https://phabricator.dev.yugabyte.com/D3616