Fix keyspace resolution for User-Defined Types
Summary:
Within a CREATE TABLE statement, the keyspace of an UDT (if not explicitly given) should resolve
to the keyspace of the table being created.
Previously, we were defaulting to the current keyspace, but Cassandra prioritizes the table
keyspace if given (since UDTs can only be used in the keyspace where they are defined).

Test Plan: jenkins, TestUserDefinedTypes.java

Reviewers: neil, robert

Reviewed By: robert

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D3635