ENG-2502: Implementing now(), totimestamp(), tounixtimestamp() and unixtimestampof() functions for YQL.
Summary:
This commit implements the following functions for YQL:

now(): This has the same functionality as now() function in CQL. It generates a new unique timeuuid. The value returned by now() is guarenteed to be unique accross space and time.

totimestamp(): This function converts a time uuid to the corresponding timestamp.

tounixtimestamp(): This function converts time uuids and timestamps to a unix timestamps, which is basically equal to the number of milliseconds since epoch.

unixtimestampof(): This function converts time uuids to a unix timestamp which is equal to the number of milliseconds since epoch.

Test Plan:
Tested by:
1) Manually checking that we can execute now() experessions in YQL and that it returns a unique Time UUID.
2) Will test it with other types by converting time uuid to timestamp types and verifying that it gives a valid representation of type.
3) Tested the MAC address bytes generated by doing the byte operations manually and checking that it is correct. Since MAC address is same for local nodes, its SHA-1 hash is also same.
4) Will have to figure out a way to unit test this.

Reviewers: robert, mikhail, pritam.damania, mihnea

Reviewed By: pritam.damania, mihnea

Subscribers: kannan, mikhail, karthik, yql

Differential Revision: https://phabricator.dev.yugabyte.com/D3676