ENG-2367: Fix for bug with AWSInitializer causing OutOfMemory error
Summary:
Long story short, we currently query for all pricing details in EC2, which results in a JSON blob that's about 180 MB. This wouldn't be an issue except that the Java JSON parser we use seems to be creating a copy of the data every time we retrieve a node within the JSON blob. This means that 180 MB quickly bloats to something like 10x it's size. One fix is to use a better Garbage Collector, but that won't necessarily work in every environment we deploy in (notably, it's unclear if it would work in containerized environments).

This fixes the issue by grabbing the pricing information only for the regions we care about and doing it one region at a time, as suggested [[ https://aws.amazon.com/blogs/aws/aws-price-list-api-update-regional-price-lists/ | here ]]. It means more network requests, but significantly smaller payloads. The pricing blob for us-west-2, for example, is only 17 MB and AWSInitializer now only takes a few seconds to run.

NOTE: This error is blocking for D3601 since I can't run itest locally and verify my changes. Also, I don't understand how itest in AWS is passing and that needs to be investigated. It should be hitting this issue.

Test Plan:
- go to UI, delete provider, recreate provider, verify that all the expected instance type data and such still appears and that the OOM error doesn't show up
- sbt test

Reviewers: bogdan

Reviewed By: bogdan

Subscribers: jenkins-bot, venkatesh, ram, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D3674