A forward Seek() should not rebuild the merge-iterator heap
Summary:
Incrementally updating the iterator tree on a Seek() instead of rebuilding the heap. This is
the way it used to work in HBase, and this might make our Next()-based workaround
unnecessary.

The downside of this change is that there is an additional key comparison to detect the case
of seeking forward. This could be avoided by introducing a new SeekForward() function that
could be called directly in case we know at the DocDB layer that the seek key is higher than
the current key.

There is no significant performance difference on the workloads I tried, with or without the
Next-instead-of-Seek workaround, but this change seems like a reasonable thing to do that
might help workloads where we have to skip over a lot of keys, i.e. reading only a subset of
columns from a wide-row table (i.e. a table with many columns).

Test Plan:
Jenkins
Performance tests using CassandraKeyValue and CassandraTimeseries read/write workloads on a single-tserver local cluster.

Reviewers: amitanand, timur, sergei, akashnil

Reviewed By: sergei, akashnil

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D3664