ENG-2469: Catalog Manager per tablet replicas state is incorrect.
Summary:
From the customer logs, there seems to be incorrect processing of heartbeats after a master becomes a new leader. One of the tservers heartbeats 3times while the sys_catalog was being loaded and was returned error correctly. And the other two heartbeat  (no proof with older logs, but that is the theory), and sneaked in to the else case for `(!l.leader_status().ok())` and were marked as "live" and "full-report-not-needed". So LB was using those live ones for AddServer purposes. This also fixes CRUD ops to not use partially added in-memory state for tablet replicas.

So the fix is to do a full fledged is-leader-ready check at the top of the `TSHeartbeat` to not allow heartbeats during leader setup to sneak in. Second change is to not unset full-report-needed in TsDesc on incremental-reports.

Misc logging to report mismatch in-memory state and persisted consensus in sys-catalog table. Some other code dedup/cleanup.

@bogdan and I tried to reproduce manually, but not luck yet. So yet to add unit tests, in the meantime, this also adds highly potential fixes and more debug info... Will be debugging/repro'ing more in followups.

Test Plan:
Jenkins

Reviewers: bogdan, pritam.damania

Reviewed By: pritam.damania

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D3677