Revert changes setting version in client
Summary:
As part of D3414.
Moved universe_version to UI payload, to fix what seemed like a bug to me.

Based on comments by @ram @bharat , that is the expected behaviour and I-Test is doing the correct thing.

Test Plan: With the changes to versioning reverted, i-test upgrade step will pass as expected.

Reviewers: jeff, ram, bharat

Reviewed By: ram, bharat

Subscribers: jenkins-bot, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D3678