ENG-2497: Changing the release versioning code
Summary:
Changed the version code to no longer generate a header full of defines. Instead, we will
prepare a JSON file that will be added to the release tar.gz, which will contain various metadata.
We will start with the already existing version pieces, such as version number, build number, type
and so on.

This diff changes the gen_version_info.py script to generate the JSON file. It also changes the
VersionInfo class to load said file from disk into memory and store it in a VersionInfoPB that
can be used across the code at runtime.

I've also updated the thirdparty rapidjson version, to get support for easily reading JSON files.

Test Plan: yb-ctl and jenkins

Reviewers: bharat, karthik, mikhail

Reviewed By: mikhail

Subscribers: ram, ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D3611