ENG-2563: Fix for backwards compatibility bug with clusters change
Summary:
In the toJson method, detect if the clusters list is empty but the userIntent/placementInfo is not. If so, move userIntent and placementInfo into a new primary cluster and add it to the universe. When any future operation is done, the json in the db will update, no need to update it right away.

Alternative would be to put together a sql update script.

Test Plan:
- sbt test
- copy over an older-style universe from dev.portal and verify the bug, then apply change and verify the bug disappears
result:
{F7813}

Reviewers: bogdan, ram

Reviewed By: ram

Subscribers: jenkins-bot, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D3687